Use Int64 for date to avoid the Year 2038 problem.

+++ DuoWeb/DuoWeb.cs
@@ -167,7 +167,7 @@ namespace Duo
 				return null;
 			}
 
-			int expire_ts = Convert.ToInt32(expire);
+			long expire_ts = Convert.ToInt64(expire);
 			if (ts >= expire_ts) {
 				return null;
 			}
